Auto Rotate

+++ src/PictureUtils/PreProcessPicture.h
@@ -20,7 +20,7 @@ void strengthenEdge(Picture *picture);
 void pushBack(Picture *picture);
 void detectEdge(Picture *picture);
 void upContrast(Picture *picture);
-Picture rotate(Picture picture, int degree);
+Picture rotate(Picture picture, float degree);
 Picture pixelsToSquare(Picture p);
 void denoizing(Picture *p);
 Pixel denoizingMedianWeight(double *m1, Pixel *m2, int w, int h, int width, int height);
